mn_rate_check test integrated with rpc-tests
