+++ qa/rpc-tests/testtools.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import sys, os
-sys.path.append('lib')
+sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), 'lib')))
 
 from decimal import Decimal
 from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
